temporal-project/worker-server/pyproject.toml
# ────────────────────────────────────────────────────────────────────────────
#TASK 2: Create Requirements File
#────────────────────────────────────────────────────────────────────────────

[project]
name = "temporal-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv>=1.0.1",
    "temporalio>=1.5.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
